Fix webpack config (#457)
